backend/requirements.txt
fastapi==0.119.0
python-decouple==3.8
uvicorn==0.37.0
boto3>=1.34.0
pydantic>=2.0.0
tensorflow>=2.15.0
numpy>=1.26.0
pillow>=10.0.0
python-multipart>=0.0.6
